+++ platform/ios/src/MGLMapView.mm
@@ -3531,9 +3531,9 @@ public:
     }
 }
 
-- (nullable NS_ARRAY_OF(id <MGLOverlay>) *)overlays
+- (nonnull NS_ARRAY_OF(id <MGLOverlay>) *)overlays
 {
-    if (self.annotations == nil) { return nil; }
+    if (self.annotations == nil) { return @[]; }
 
     NS_MUTABLE_ARRAY_OF(id <MGLOverlay>) *mutableOverlays = [NSMutableArray new];